* main: (110 commits)
  Escape a character causing a doxygen error.
  Updated release notes.
  Added a comment block for future reference.
  more syntax fixes
  Update CMakeLists.txt
  CMake: Find HDF5 header we can safely include for other checks
  moving functions and macros to new file, lowercase things
  Update release notes.
  lowercase
  lower case
  lowercase
  moving functions and macros to a file
  moving the dependencies inclusion
  CMake: Add support for UNITY_BUILD
  removing debug messages
  actually adding the dependencies file...
  putting dependencies into separate file
  Define USE_SZIP variable for nc-config.cmake.in
  matching cmake variables in autotools configuration
  moving the version into the project command in cmake
  ...

+++ h5_test/tst_h_vl.c
@@ -7,6 +7,7 @@
 
 #include "h5_err_macros.h"
 #include <hdf5.h>
+#include <stddef.h>
 
 #define FILE_NAME "tst_h_vl.h5"
 #define DIM1_LEN 3
@@ -23,7 +24,7 @@ main()
       hsize_t dims[1] = {DIM1_LEN};
       hvl_t data[DIM1_LEN], data_in[DIM1_LEN];
       int *phoney;
-      int i, j;
+      size_t i, j;
       size_t size;
       
       /* Create some phoney data, an array of struct s1, which holds a
